Allow disabling the pthread support check
In cases where this code is utlized only for the libriers in it
(e.g. vTPM emulation in Coconut SVSM), we do not need to compile
the simulator.

So it would be convenient to have the ability to disable pthread
checking, which in that case is not necassary since it is used only by
the simulator. Also in Coconut SVSM we don't have pthread support, so
it's also a requirement and for now we have our own fork to disable
this check.

By default pthread checking remains enabled, so there is no change.
`--disable-pthread` is added to the configure to support this new use
case.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>